Fix memory issues
Properly initialize df_changes to avoid uninitialize memory crash

Also, properly initialize df_lnum / df_count to clear them to zero.
We could make sure to never use the invalid values if the buffer is not
NULL but in practice that is quite tricky and is prone to error making
it easy to do memory bugs. Just set everything to 0 to make it easy to
reason about.