runtime(netrw): Fix directory netrw-mc error on unix-like systems
When attempting to copy a marked directory to a target directory the
g:netrw_localcopydircmdopt was ignored resulting on unix systems to
attempt to use the `cp` command without the required `-R` argument.

This change captures the g:netrw_localcopydircmdopt value and appends it
to the system() call when the marked file(s) is a directory.

Signed-off-by: Devin Weaver <suki@tritarget.org>