Fix UAF when switching to a closing buffer
Problem: Possible to open more windows into a closing buffer without splitting,
bypassing existing "b_locked_split" checks and triggering UAF.

Solution: Disallow switching to a closing buffer. Editing a closing buffer (via
":edit", etc.) was fixed in v9.1.0764, but add an error message and check just
"b_locked_split", as "b_locked" is necessary only when the buffer shouldn't be
wiped, and may be set for buffers that are in-use but not actually closing.