Fix 'eventignorewin' for non-current buffers
Problem: When an autocommand executes for a non-current buffer, 'eventignorewin'
is only checked from the buffer's last wininfo (overwrites win_ignore in the
loop), not from the value of 'eventignorewin' in all windows showing the buffer
as described.

Solution: Fix the check and don't use wininfo, as that may only contain windows
that recently showed the buffer. Consider all the buffer's windows in all
tabpages.