Fix memory leak when buflist_new fails to reuse curbuf
Problem: buflist_new leaks ffname and fails to reuse curbuf when autocommands
from buf_freeall change curbuf. Plus, a new buffer is not allocated in this
case, despite what the comment above claims.

Solution: Remove the condition so ffname is not leaked and so a new buffer is
allocated like before v8.2.4791. It should not be possible for undo_ftplugin or
buf_freeall autocommands to delete the buffer as they set b_locked, but to stay
consistent with other uses of buf_freeall, guard against that anyway.

Note that buf is set to NULL if it was deleted to guard against the (rare)
possibility of messing up the "buf != curbuf" condition below if a new buffer
happens to be allocated at the same address.