docs(pi_tutor): remove outdated Contribution section
Problem:
The Github repo link in the Contribution section has been archived for 5
years. So people who want to contribute to the Tutor plugin should just
send PR to Vim repo, similar to most other Vim features, so there is no
need for a Contribution section in the plugin doc.

Solution:
Replace it with an Original Author note in the beginning of the document