Fix :mark arguments validation
Problem:  :mark can be used to set marks [0-9] despite the documentation
          explicitly declaring that this is not possible.
Solution: Output an "Invalid argument" error when [0-9] is specified as
          an argument to :mark or :k.

Update the documentation at :help :mark to reflect that the [, ], <, and
> marks can also be set with this command.  It appears the documentation
was not updated when m[, m], m< and m> were added.

Replace the overly specific "E191:" error, in favour of the more general
"E475: Invalid argument" used by most commands, rather than update it to
reflect all allowed arguments.

Signed-off-by: Doug Kearns <dougkearns@gmail.com>