Fix showbreak and cursorcolumn
This changes the behaviour of how the showbreak option is treated in
combination with color column. Previously the color column could also be
displayed on the "virtual text" of the showbreak text.

As this is not what the color column is supposed to be, the color column
now skips the showbreak columns and only displays in actual text or
after the actual text.