CI: Github Runner TestGUI consistently fails
It's unclear to me why exactly one testgui runner fails.

It seems there the runner run the various screen tests when the GUI is
active with a slight different window size (although all my debugging
attempts could not verify this thesis).

So in order to keep fix the runners, let's disable a few tests for the
Github runners when GUI mode is active.

Also, run the tests without a toolbar, it throws a few GTK related
warnings such as this:

```
(gvim:19972): Gtk-WARNING **: 21:17:50.412: GtkToolbar 0x5628d7460f70
reported min size 43 and natural size 7 in get_preferred_width();
natural size must be >= min size
```
which may or may not be related to the failure.

fixes: #17595

Signed-off-by: Christian Brabandt <cb@256bit.org>