Add a generic error when diff anchors fail to resolve at diff time
This gives a general error to let the user know the anchor resolution
has failed, in addition to whatever specific message they see like
"invalid mark". Also, currently a failed pattern search does not show
the error message since we are searching with silent mode on, so this is
the only error they will see.