Make get_address() not add search pattern to @/ register
This prevents diff anchors using pattern search from polluting @/ and to
make sure it doesn't reset the search highlight. The only other user of
this function which relies on the silent flag is the incsearch but that
function already caches the pattern using `save_last_search_pattern()`.
We could do that for diff anchors too but it's simply more efficient to
just not pollute the register to begin with. It's unlikely a caller
passing "silent" in would want to do that.