feat: enable colored caption bar on Windows
Problem: gVim on Windows' caption bar is always white.

Solution: add gui_mch_set_caption() to cope with that.

The implement of 'guidarkmode' is a much longer task, so I would like to
complete this most needed feature.

This commit seen the caption bar as the extension of the background so
it is follow the option 'background' before Windows 11 and follow the
background color after Windows 11.

Closes: #3922, #18028

Signed-off-by: Mao-Yining <mao.yining@outlook.com>