patch 9.1.1813: MS-Windows: title bar is always white
Problem:  MS-Windows: title bar is always white
Solution: Set_caption() and set the title bars color to the
          Normal highlighting background (Mao-Yining)

The implement of 'guidarkmode' is a much longer task, so I would like to
complete this most needed feature.

This commit seen the caption bar as the extension of the background so
it is follow the option 'background' before Windows 11 and follow the
background color after Windows 11.

fixes: #3922
fixes: #18028
closes: #18282

Signed-off-by: Mao-Yining <mao.yining@outlook.com>
Signed-off-by: Christian Brabandt <cb@256bit.org>